feat(Filenames): Capitalizing all component file names
