+++ src/EventHandler.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import Observable from 'ol/Observable';
-import EventHandler from './event-handler';
+import EventHandler from './EventHandler';
 
 
 describe('<EventHandler />', ()=> {
