+++ src/View.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import OlView from 'ol/View';
 import {consumer} from './hoc';
-import EventHandler from './event-handler';
-import {MapCtx} from './map';
+import EventHandler from './EventHandler';
+import {MapCtx} from './Map';
 
 
 const updateView = (view, {center, zoom, rotation, minZoom, maxZoom})=> {
