+++ src/layer/Group.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import LayerGroup, {LayerGroupCtx} from './group';
+import LayerGroup, {LayerGroupCtx} from './Group';
 
 
 describe('<LayerGroup />', ()=> {
