+++ src/layer/Tile.js
@@ -1,32 +1,32 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import OlVectorLayer from 'ol/layer/Vector';
+import OlTileLayer from 'ol/layer/Tile';
+import EventHandler from '../EventHandler';
 import {consumer} from '../hoc';
 import {LayerCtx} from '.';
-import {LayerGroupCtx} from './group';
+import {LayerGroupCtx} from './Group';
 
 
 @consumer(LayerGroupCtx)
-class VectorLayer extends React.Component {
+class TileLayer extends React.PureComponent {
   static propTypes = {
     layerGroup: PropTypes.object,
     id: PropTypes.string,
     visible: PropTypes.bool,
-    style: PropTypes.func,
-    children: PropTypes.any
+    children: PropTypes.any,
+    onPreCompose: PropTypes.func,
+    onPostCompose: PropTypes.func
   };
 
   constructor(props) {
     super(props);
 
-    this.layer = new OlVectorLayer({
-      id: this.props.id,
-      updateWhileAnimating: true,
-      updateWhileInteracting: true
+    this.layer = new OlTileLayer({
+      id: this.props.id
     });
   }
 
-  // TODO: should this live in constructor
+  // TODO: should that live in the constructor
   componentDidMount() {
     const {layerGroup} = this.props;
     layerGroup.getLayers().push(this.layer);
@@ -40,20 +40,26 @@ class VectorLayer extends React.Component {
 
   componentDidUpdate() {
     const {visible} = this.props;
+    const {layer} = this;
 
-    this.layer.setVisible(visible);
+    layer.setVisible(visible);
   }
 
   render() {
     const {layer, props} = this;
+    const {children, onPreCompose, onPostCompose} = props;
 
     return (
       <LayerCtx.Provider value={{layer}}>
-        {props.children}
+        <EventHandler
+          target={layer}
+          precompose={onPreCompose}
+          postcompose={onPostCompose}
+        />
+        {children}
       </LayerCtx.Provider>
     );
   }
 }
 
-
-export default VectorLayer;
+export default TileLayer;
